// src/store/types/emailReceiver.actionTypes.ts
export enum EMAIL_RECEIVER_ACTION_TYPES {
  // Start Polling
  START_POLLING_REQUEST = 'START_POLLING_REQUEST',
  START_POLLING_SUCCESS = 'START_POLLING_SUCCESS',
  START_POLLING_FAILURE = 'START_POLLING_FAILURE',

  // Stop Polling
  STOP_POLLING_REQUEST = 'STOP_POLLING_REQUEST',
  STOP_POLLING_SUCCESS = 'STOP_POLLING_SUCCESS',
  STOP_POLLING_FAILURE = 'STOP_POLLING_FAILURE',

  // Check Now
  CHECK_NOW_REQUEST = 'CHECK_NOW_REQUEST',
  CHECK_NOW_SUCCESS = 'CHECK_NOW_SUCCESS',
  CHECK_NOW_FAILURE = 'CHECK_NOW_FAILURE',

  // Get Status
  GET_STATUS_REQUEST = 'GET_STATUS_REQUEST',
  GET_STATUS_SUCCESS = 'GET_STATUS_SUCCESS',
  GET_STATUS_FAILURE = 'GET_STATUS_FAILURE',

  // Clear Errors
  CLEAR_EMAIL_RECEIVER_ERRORS = 'CLEAR_EMAIL_RECEIVER_ERRORS',
}